+++ packages/helpers/src/jora/index.ts
@@ -11,6 +11,7 @@ import semverParse from 'semver/functions/parse';
 import semverSatisfies from 'semver/functions/satisfies';
 import { Range, SemVer } from 'semver';
 import networkTypeList, { bytesInMBit, Item } from '../network-type-list';
+import Graph, { Node as GraphNode, PathSolution } from '../graph';
 import { colorFromH, colorMap, fileTypeMap, generateColor } from './colors';
 import { pluralEng, pluralRus } from './plural';
 
@@ -201,6 +202,23 @@ export default function helpers() {
     },
     isMatch(a: string, b?: string | RegExp): boolean {
       return b instanceof RegExp ? b.test(a) : typeof b === 'string' ? a === b : true;
+    },
+
+    graph_getNode<TData>(id?: string, graph?: Graph<TData>): GraphNode<TData> | null {
+      return graph?.getNode(id!) ?? null;
+    },
+
+    graph_getPaths<TData>(
+      from?: GraphNode<TData>,
+      graph?: Graph<TData>,
+      to?: GraphNode<TData>,
+      max = Infinity
+    ): PathSolution<TData> | null {
+      if (!from || !to || !graph) {
+        return null;
+      }
+
+      return graph.findPaths(from, to, max);
     },
   };