+++ packages/webpack-plugin/src/index.ts
@@ -13,7 +13,6 @@ import { Extension } from '@statoscope/stats/spec/extension';
 import WebpackCompressedExtension from '@statoscope/webpack-stats-extension-compressed';
 import WebpackPackageInfoExtension from '@statoscope/webpack-stats-extension-package-info';
 import { CompressFunction } from '@statoscope/stats-extension-compressed/dist/generator';
-import Piper from '@statoscope/report-writer/dist/piper';
 
 export type Options = {
   name?: string;
@@ -93,7 +92,7 @@ export default class StatoscopeWebpackPlugin {
         statoscopeMeta.extensions.push(compressedExtension.get());
       }
 
-      const webpackStatsStream = new Piper(stringifyStream(statsObj) as Readable);
+      const webpackStatsStream = stringifyStream(statsObj);
       let statsFileOutputStream: Writable | undefined;
       let resolvedSaveStatsTo: string | undefined;
 
@@ -103,12 +102,12 @@ export default class StatoscopeWebpackPlugin {
         );
         fs.mkdirSync(path.dirname(resolvedSaveStatsTo), { recursive: true });
         statsFileOutputStream = fs.createWriteStream(resolvedSaveStatsTo);
-        webpackStatsStream.addConsumer(statsFileOutputStream);
+        webpackStatsStream.pipe(statsFileOutputStream);
       }
 
       const statsForReport = this.getStatsForHTMLReport({
         filename: resolvedSaveStatsTo,
-        stream: webpackStatsStream.getInput(),
+        stream: stringifyStream(statsObj),
       });
       const htmlReportPath = this.getHTMLReportPath();
       const resolvedHTMLReportPath = path.resolve(
@@ -119,7 +118,7 @@ export default class StatoscopeWebpackPlugin {
       try {
         await Promise.all([
           htmlReport.writer?.write(),
-          webpackStatsStream.consume(),
+          waitStreamEnd(statsFileOutputStream),
           waitStreamEnd(htmlReport.stream),
         ]);